backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
transformers>=4.35.0
torch
pydantic>=2.4.0
python-multipart
psutil
python-dotenv
